+++ web/src/types/input.ts
@@ -5,4 +5,5 @@ export type InputProps = {
   value: string | number;
   onChange: (e: React.ChangeEvent<HTMLInputElement>) => void;
   placeholder?: string;
+  checked?: boolean;
 };
